Refactored according to teacher's comment
